Add Authentication for Label page
