+++ public/html/index.html
@@ -11,7 +11,7 @@
 </head>
 <body>
     <nav class="navbar navbar-expand-lg navbar-light border-bottom"> <!--  style="background-color: #e3f2fd;"-->
-        <a href="#" class="navbar-brand">Label Me</a>
+        <a href="/" class="navbar-brand">Label Me</a>
 
         <form class="col-12 col-lg-auto mb-3 mb-lg-0 me-lg-3 ml-auto">
             <input type="search" class="form-control form-control-dark" placeholder="Search..." aria-label="Search">
@@ -24,7 +24,8 @@
         <div id="navbarCollapse" class="collapse navbar-collapse">
             
             <ul class="navbar-nav ml-auto">
-                <li class="nav-item active"><a href="./html/welcome.html" class="nav-link">Login</a></li>
+                <li id="nav-login" class="nav-item"><a href="./html/welcome.html" class="nav-link">Login</a></li>
+                <li id="nav-label" class="nav-item"><a href="./html/label.html" class="nav-link">Label</a></li>
                 <!-- <li class="nav-item"><a href="#" class="nav-link">profile</a></li> -->
                 <li class="nav-item">
                     <a href="#" class="d-block link-dark text-decoration-none dropdown-toggle" id="dropdownUser1" data-bs-toggle="dropdown" aria-expanded="false">
@@ -90,6 +91,6 @@
             </p>
         </div>
     </footer>
-
+    <script src="../js/index.js"></script>
 </body>
 </html>
