+++ public/html/label.html
@@ -11,9 +11,13 @@
     <!-- <script src="https://cdnjs.cloudflare.com/ajax/libs/fabric.js/4.4.0/fabric.min.js"></script> -->
 </head>
 <body>
-    <nav class="navbar navbar-expand-lg navbar-light" style="background-color: #e3f2fd;">
-        <a href="#" class="navbar-brand">Label Me</a>
+    <nav class="navbar navbar-expand-lg navbar-light border-bottom"> <!--  style="background-color: #e3f2fd;"-->
+        <a href="/" class="navbar-brand">Label Me</a>
 
+        <form class="col-12 col-lg-auto mb-3 mb-lg-0 me-lg-3 ml-auto">
+            <input type="search" class="form-control form-control-dark" placeholder="Search..." aria-label="Search">
+        </form>
+        
         <button class="navbar-toggler" type="button" data-toggle="collapse" data-target="#navbarCollapse">
             <span class="navbar-toggle-icon"></span>
         </button>
@@ -21,17 +25,17 @@
         <div id="navbarCollapse" class="collapse navbar-collapse">
             
             <ul class="navbar-nav ml-auto">
-                <li class="nav-item active"><a href="#" class="nav-link">Login</a></li>
-                <li class="nav-item"><a href="#" class="nav-link">Sign up</a></li>
-                <li class="nav-item"><a href="#" class="nav-link">User profile</a></li>
+                <li class="nav-item">
+                    <a href="#" class="d-block link-dark text-decoration-none dropdown-toggle" id="dropdownUser1" data-bs-toggle="dropdown" aria-expanded="false">
+                        <img src="../images/member.png" alt="mdo" width="32" height="32" class="rounded-circle">
+                    </a>
+                </li>
             </ul>
         </div>
 
     </nav>
-
     <div class="container">
         <div class="pt-4 bg-dark text-white my-3 border">
-            <!-- <h1>Label ME</h1> -->
             <p>1. Choose your image and then upload it.</p>
             <p>2. Use annotation tools to label!</p>
             <p>3. Make sure to commit each steps to save your labeling.</p>
@@ -88,8 +92,17 @@
 
         <canvas id="canvas" width="800" height="600" style='border:1px solid #000000;'></canvas>
     </div>
+    <footer class="footer mt-auto text-muted py-5" style="background-color: #e3f2fd;">
+        <div class="container">
+            <p class="float-end mb-1">
+                <a href="#">Back to top</a>
+            </p>
+            <p class="mb-1">
+                LabelMe © Bootstrap
+            </p>
+        </div>
+    </footer>
     <script src="../js/fabric.min.js"></script>
-    <script src="../js/index.js"></script>
-
+    <script src="../js/label.js"></script>
 </body>
 </html>
